backEnd/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.335
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai